% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expSCI.R
\name{expSCI}
\alias{expSCI}
\title{Find simultaneous confidence for the successive pair-wise differences of exponential locations}
\usage{
expSCI(data, alpha)
}
\arguments{
\item{data}{a real matrix}

\item{alpha}{real number between 0 and 1 called significance level}
}
\value{
numeric vector
}
\description{
More detailed description
}
\examples{
k=3;N=c(20,30,20);S=c(1,1,1);alpha=0.05
g=NULL
for(i in 1:k)
{
 g[[i]]=rexp(N[i],1/S[i])
}
data=g
expSCI(data, 0.05)
}
