% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exptest.R
\docType{package}
\name{Exptest}
\alias{Exptest}
\title{Find test statistic value from a given data and corresponding critical value}
\description{
This package finds test statistic value and
critical value of PB tests for testing equality of locatio parameters of exponential
distributions against
ordered alternatives with heteroscedasticity.
}
\author{
Anjana Mondal \email{mondal.anjana25@gmail.com}
}
