% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxT.R
\name{MaxT}
\alias{MaxT}
\title{Find test statistic value from a given data and corresponding critical value}
\usage{
MaxT(data, alpha)
}
\arguments{
\item{data}{a real matrix}

\item{alpha}{real number between 0 and 1 called significance level}
}
\value{
numeric vector
}
\description{
More detailed description
}
\examples{
k=3;N=c(20,30,20);S=c(1,1,1)
g=NULL
for(i in 1:k)
{
 g[[i]]=rexp(N[i],1/S[i])
}
data=g
MaxT(data, 0.05)
}
